feat: Refactor ProfileAvatar and App components to use centralized profile fetching from the store
